[github] init: Add clean architecture flow with reactive programming
This commit is the last "init" related commit and it is the one that
initialise the first clean architecture flow, while utilising RxJava.

This change consist of:
1) Adding the RxJava and RxAndroid dependency to the project.
2) Adding all the Rx related helper classes and dependencies that will
   compose our first abstract use case (a UseCaseSingle).
3) Adding our first use case (the SearchUseCase).
4) Adding our first repository (the SearchRepository) and its
   related implementing class.
5) Giving life to our first view model class that with the help of the
   corresponding use case will trigger the search.
6) Finally, giving life to our first activity observer, which will
   observer the view model for changes and update the UI accordingly.

From now on and for every new feature, this clean architecture pattern
will be utilised, along-side the MVVM pattern (using the architecture
component ViewModel class).