[github] test: Add test module
This "test" module is a helper module that will provide domain models
to test classes. This is done in order to remove duplication from tests
that need to create a domain model or other objects repetitively.

Also, the helper functions make sure to provide the developers the
means to instantiate a default domain model object, without providing
any extra information. Still, those functions make it easy for the
developers to use argument parameters if they need a domain model other
than the default.

In general, this patterns helps reduce the setup boilerplate from tests
and the duplication that creates.